+++ CHANGELOG.md
@@ -27,6 +27,8 @@ Changelog
 -   Support non-gate instructions (e.g. `MEASURE`) in `to_latex()` (@notmgsk, gh-975).
 -   Test suite has been updated to reduce the use of deprecated features
     (@kilimanjaro, gh-998).
+-   Certain tests have been marked as "slow", and are skipped unless
+    the `--runslow` option is specified for `pytest` (@kilimanjaro, gh-1001).
 
 ### Bugfixes